backend-scm/requirements.txt
# SCM-INSIGHTS Backend - same stack as main backend
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
bcrypt==4.0.1
python-dotenv==1.0.0
psycopg[binary,pool]>=3.1
requests==2.31.0
